backend/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
atomicwrites==1.4.1
attrs==22.1.0
beautifulsoup4==4.12.2
binary==1.0.0
cachetools==4.2.2
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.1
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.16
django-cors-headers==3.14.0
django-debug-toolbar==2.2
django-extensions==3.2.3
django-extra-fields==3.0.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-extra-fields==3.7.0
exceptiongroup==1.1.0
Faker==12.0.1
filetype==1.2.0
flake8==6.0.0
flake8-docstrings==1.6.0
flake8-isort==6.0.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.1.2
more-itertools==8.2.0
oauthlib==3.2.2
packaging==21.3
Pillow==9.3.0
pluggy==0.13.1
psycopg==3.1.12
psycopg2==2.9.7
psycopg2-binary==2.9.7
py==1.11.0
pyaes==1.6.1
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
PyJWT==2.1.0
pyparsing==3.0.9
Pyrogram==2.0.103
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.3
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
six==1.14.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sorl-thumbnail==12.6.3
soupsieve==2.4.1
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.0
tornado==6.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.25.6
wcwidth==0.1.8
webcolors==1.11.1
zipp==2.2.0
